Improve DelegatingConstraintCheckerTest
Instead of a JsonFileEntityLookup, we use an InMemoryEntityLookup and
create the items in the tests that use them, which makes the tests much
more readable and lets us get rid of some bloated JSON files. The tests
also use checkAgainstConstraintsOnEntityId instead of
checkAgainstConstraints, because I plan to remove the latter method
(it’s almost entirely redundant with checkAgainstConstraintsOnEntityId).

Change-Id: I4c6528fb38de33562770455b794dea6620638907