Convert quantities to standard units
When computing a comparison or a difference between two quantities,
attempt to convert each of them to standard units first.

The tests use SI unit symbols instead of Q-IDs, since nothing in
Wikibase requires Q-IDs and it makes the tests more readable.

The configuration setting for the “year” unit is not removed, even
though it is currently unused: a later commit will add support for
calculating time differences in years as well, since years cannot be
converted to seconds and several “difference within range” constraints
on Wikidata have a range expressed in years. (Using the average length
of a Gregorian year – 365.2425 days – would yield the undesirable result
that the difference between November 12, 1955 and November 12, 1956 is
slightly more than one year (1956 is a leap year), leading to a
confusing constraint violation for a range of [0 years, 1 year].)

Bug: T164371
Change-Id: Ic5dee62c013db71a16a306f564b15baa50fee098