Check “unique value” constraint on qualifiers and references
The constraint requires that a qualifier value be unique across all
other qualifiers, or that a reference value be unique across all other
reference snaks, but does not consider a value occurring in a qualifier
and another reference to be a violation.

For deprecated statements, the reference snaks are still checked but the
qualifier snaks aren’t, since the qualifiers are part of the claim
(which the rank applies to) but the references aren’t. This strategy
might change if we reach a different decision in T171762.

The implementation is a bit tricky because unlike the entire statement,
the qualifiers and references don’t have an identifier of their own, so
to look for them, we actually have to serialize the value into the
SPARQL query. For value types with a full value representation, we can
use the full value node (wdv:$hash) as shortcut; for other value types,
we reimplement a bit of Wikibase’s RDF serialization code.

Bug: T175561
Change-Id: Ic997019ab7651fc11365c895d01ca2fd70e1ca7d