Finish new check result status for T175570
In Ia82a80aeb3, I added a new check result status for when a constraint
check is skipped because the context is not a 'statement'-type context.
This commit cleans it up a bit:

1. Change it to the slightly clearer 'not-main-snak'.
2. Add a missing i18n/ message, which is used on the special page.
3. Add a missing color, also used on the special page.
4. Explicitly declare its sort order. (Also, use named constants there.)

Also, document steps 2, 3 and 4 below the last status in CheckResult.php
so that perhaps they’re not forgotten the next time a status is added.

Bug: T175570
Change-Id: Ia3a14e4e232b2ceee4b100ccd07b177fc8f870ad