Fix CommonsLinkChecker for Creator namespace
The TitleParser we use is configured with the namespaces of the local
wiki (on Wikidata, wikidatawiki), which are not the same as the
namespaces of Wikimedia commons – for instance, the Creator namespace is
commonswiki-specific (namespace number 100). Due to the way Wikimedia’s
MediaWiki config is set up, there’s no proper way to fix this, so we
just hard-code the few commonswiki-specific namespace numbers.

We now use the TitleParser for two purposes, depending on namespace: if
we know the namespace (e. g. Creator), we just use it to turn the title
into the corresponding database key, otherwise we also have it resolve
the namespace. In both cases, TitleParser processes interwiki prefix,
namespace and proper title, which is more than we want (e. g. for a
hypothetical creator “File:Foo” – full title “Creator:File:Foo” – we
would check in the “File” namespace again), but until TitleParser is
refactored to publicly expose just the parts of its functionality we
need, I see no way around this problem.

Bug: T176062
Change-Id: Ib54cc1a42acc569e4c433e9f793e3c945cf6a549