Use per-regex cache map to cache regex check results
Instead of caching each regex check result under an individual,
per-regex and per-text key, we cache an array for each regex, where we
cache results for the least recently used texts. The size of that cache
map is configurable with the WBQualityConstraintsFormatCacheMapSize
option, defaulting to 100 entries.

The test for the cache is removed again, since refresh behavior is
nondeterministic.

Bug: T173696
Change-Id: I5e9fff029010551736c2733dcb93d9b33b110381
Depends-On: I1cd98ea8965b51e43371efd990fb9302bb507928