Add support for qualifier/reference reports to gadget
If the wbcheckconstraints response contains reports for qualifier and/or
reference snaks, the gadget extracts and displays them.

The $overlay specification in the gadget’s popup is necessary because
there is an `overflow: hidden` rule for the entire reference view (added
in Wikibase commit d2f2785c44, change I9b1323a89d), which means that the
popup’s content is hidden as well. We fix that by moving the popup in
the DOM to append it to the entire statement instead, which it can
overflow without problems. To make this possible, the buildPopup
function has to know the container, so it’s refactored to do the
appending to the container itself instead of returning the widget.

Bug: T168532
Change-Id: I1cb1103f36b231308f7bfb7c5ded02855924be81