Fix constraint checks on saving statement
In commit 4b893c1cb1 (change I99379a96cd), I changed the calling
convention of addReportsToStatement: the second argument should now
point to the whole statement, not to the main snak value. However, I
forgot to update the call in the API response handler when checking
constraints for a single saved statement (wikibase.statement.saved
hook), so the results in that case could not be attached.

Bug: T176602
Change-Id: I19f06af6c8a3fbb2ee314a31b4946673f7328b2c