Add support for “year” ranges
RangeCheckerHelper gains support for calculating the difference between
two TimeValues in years, instead of in seconds (SI base unit). Sub-year
differences are not calculated accurately, but dates within the year are
respected: the difference between March and April of the same year is
reported as 0.5 years, which is more than 0 years but less than 1 year.

ConstraintParameterParser reports an error if exactly one range endpoint
has the unit “year” and the other one has a different unit. Since years
can’t be converted to another standard time unit, we cannot compare
against such a range.

DiffWithinRangeChecker checks if the range was specified in years, and
if so, has RangeCheckerHelper calculate the difference in years as well
instead of in seconds. This requires injecting a Config into
DiffWithinRangeChecker so that it knows the “year” unit.

Taken together, this adds support for “year” ranges to “difference
within range” constraints, e. g. “difference between date of birth and
date of death must be within 0 to 150 years”. Without the extra
date-within-the-year handling, this would fail to reject items with date
of birth after the date of death, but within the same year. And if years
were simply converted to 365.2425 days (the average length of a year in
the Gregorian calendar), then an item with date of birth January 1st,
1820 and date of death 150 years later to the day would incorrectly be
reported as violation because the difference of 54787 days is slightly
more than 150×365.2425 days = 54786.3750 days.

Bug: T164371
Change-Id: I1b6bdc1a4731e3a98d9de0b0086eae504204b196