Add UsedForValuesOnly to DelegatingConstraintCheckerTest
DelegatingConstraintCheckerTest should contain at least one constraint
of every constraint type, to make sure that ConstraintReportFactory
isn’t missing any checkers when initializing the checker map.

The test is also refactored to remove the magic number of constraints
for P1 from three places where it was previously hard-coded, so that in
the future adding just an array entry is enough to add another
constraint type.

Change-Id: I7977221f86833147466da554fe930536567c02e6