Use WikiPage to insert pages in tests
Directly manipulating the page table is fragile, and has been broken on
Travis’ sqlite builds since the build for 38bb16a0ef [1]. Instead, let’s
properly create a WikiPage object and insert it into the table.

The proper way to do this would be to use doEditContent() (or insertOn()
with several follow-up method calls), but I found that to be
prohibitively slow (test time on my system went from 4s to 6s – a 50%
increase!), so we’re skipping that and inserting pages without
revisions or content.

[1]: https://travis-ci.org/wikimedia/mediawiki-extensions-WikibaseQualityConstraints/builds/283614690

Change-Id: I7756ac088df6dd9c1ad7ff39eda9d618e5d5ad36