Always include main snak context in API response
Context::storeCheckResultInArray is expanded to populate the container,
but not store anything, if the $result to store is null. A NullResult
subclass of CheckResult is added, and the CheckConstraints API maps that
to a null $result to insert into the response container, with the effect
that NullResults ensure that the structure for the corresponding context
is present in the API response even if there are no constraints defined
for that context.

Several DelegatingConstraintChecker methods gain an additional
parameter, $defaultResults, which can be used to populate the check
results beyond the normal results of the constraint checking process.
The CheckConstraints API uses this to add NullResults for statement
contexts (ensuring that main snak contexts are always present in the API
response, fixing T178160), while other contexts are left alone
(qualifiers and references may be completely absent from the response).

Bug: T178160
Change-Id: I60ad838dda6177eaec7383e898645f700e96dc96