Fix sorting of NullResults
Commit a4c1832358 (change I60ad838dda) introduced the NullResult type,
which is a fake result holding no constraint to ensure that certain
contexts still appear in the API response even if they have no
constraints defined on them. However, the DelegatingConstraintChecker
sometimes look at the constraint of a result: when sorting the results,
it first sorts by status, then by snak, but if those are identical, then
it looks at the constraint type. If there are two identical statements
on an entity included in a constraint check, they will both add a
NullResult to the result list, and the comparer will blow up trying to
compare the two NullResults, since the snaks are identical.

Bug: T179038
Change-Id: I48ac2d947020f4dbf3c6c87ec5c098a4114de373