Change stylesheet files to Less
For now, we mostly utilize the ability to import colors instead of
hard-coding their hex values, but there are probably further
improvements down the line.

Some colors that SpecialConstraintReportPage.less uses are not part of
mediawiki.ui/variables, and it’s not clear how to import the stylesheet
that does define them, so for now those remain hard-coded.

Change-Id: I5d53b985ddbfce5f3964fc5d5e32a8c1527a6596