Add function to get cache max-age in SparqlHelper
This function can then later be used to pass that max-age up the chain,
eventually to tell the user how outdated a cached constraint check
result might be.

This also adds tests, including for the new code added in the previous
commit (which this commit moves to the new function).

Change-Id: Ife6e593b10cb78a1b3bc47cd097d22aaf4b48cb5