Add CachingMetadata value object
A CachingMetadata object encapsulates information about how a value was
cached, without any information about the value itself. It can be
bundled together with a primitive value, or added to a CheckResult. It
is also possible to merge several pieces of information into one, for
results that are derived from more than one other result.

Bug: T179844
Change-Id: I27f19ef3b97618700435dacedab783df5e08461e