Add CachedArray, CachedQueryResults containers
CachedArray is simply a wrapper around an array and some
CachingMetadata, and CachedQueryResults is a functionally identical
class with some extra documentation.

SparqlHelper::runQuery() is changed to return CachedQueryResults and
attaches the caching information it extracted from the HTTP headers to
them. The other SparqlHelper methods currently throw this information
away again, but will be updated to return another container (e. g.
CachedBoolean) in later commits.

Bug: T179844
Change-Id: I80930ee94a3bef26b07f001ce2d4f346a3b6ff9a