Add CachedBool container and use in helpers
CachedBool is simply a wrapper around a bool and some CachingMetadata,
analogous to CachedArray. Several methods in SparqlHelper and
TypeCheckerHelper are updated to return a CachedBool (propagating the
CachingMetadata from the underlying runQuery() call). The information is
still thrown away eventually (this time by the constraint checkers
directly), but later commits will fix that.

Bug: T179844
Change-Id: Ie0510d7e36b5ea9e4350160fe4ca25c2293df3b3