Add HTTP header indicating gadget to API requests
This adds an HTTP header “X-MediaWiki-Gadget” to all API requests by the
gadget. If the gadget is loaded as a ResourceLoader module (as it should
be), the header’s value is “checkConstraints”, otherwise (for instance,
if someone copied the code to their common.js) the value is
“checkConstraints-custom”. This should give us an idea of how many
external users of the constraint APIs outside of the gadget exist
(corresponding to API requests without the header).

Bug: T180780
Change-Id: I37caea4c4eda720a16ab4735355dfaca51c2371a