Bring SQL file closer to MediaWiki style
Use lowercase for types, write CREATE INDEX statements in one line, and
don’t use vertical alignment.

This commit does not address some elements of the table structure itself
that are also not in line with MediaWiki style:

- Table names are usually singular nouns, so it should be
  “wbqc_constraint”, not “wbqc_constraints”.
- The “pid” column name should have the same prefix as all the other
  columns, so it should be “constraint_pid”.
- Arguably, the “constraint_” prefix is also too long.

We may address these at some point, but the process to change the schema
is much more complicated than simply reformatting the SQL file.

Change-Id: I20618527155d0a75220a6c859c0640a15dbcb339