Add tests for WikibaseQualityConstraintsHooks::isConstraintStatementsChange
To make the tests possible, the method is slightly refactored to get its
Config injected (so the tests don’t use the real installation’s config),
and it’s also made public so the tests can access it.

Most of the negative test cases are taken from Wikibase Lib’s own tests,
but some negative test cases and all of the positive test cases have to
be constructed more manually. (But at least we can still use the given
EntityChangeFactory.)

Bug: T182110
Change-Id: Icd6182bded426617ef401f19a801a9df3ac7fcf3