Implicitly track property of constraint
For now, we also want to invalidate cached check results whenever the
property is changed, because the change might have edited the constraint
statements. We might be able to improve this behavior in the future,
since with T182110 we know which edits changed the constraint statements
and which didn’t, but for now invalidating on any property edit is also
acceptable since properties aren’t edited too frequently anyways.

Bug: T182104
Change-Id: I5ce1ef3222a17be1f6238631de7591e330023883