Add and use CachedCheckConstraintsResponse
CachedCheckConstraintsResponse is, like CachedQueryResults and
CachedEntityIds, a functionally identical subclass of CachedArray with
some extra documentation. ResultsBuilder is changed to return that class
instead of a bare array, so that another kind of ResultsBuilder (to be
added in a later commit) can inspect the list of depended entity IDs
when caching the entire response.

Bug: T182105
Change-Id: I1433ea7e32a269bcf322a55d56d85bdce1cb574a