Add and use CachingResultsBuilder
This ResultsBuilder implementation wraps another ResultsBuilder and
stores its result in some cache, along with some information to be able
to determine whether a cached result is fresh or not. For now, it does
not retrieve the result from that cache – that will be implemented later
(but that’s the reason why getAndStoreResults is a separate method and
not inlined into getResults).

CachingResultsBuilder is only used in the CheckConstraints API if the
config setting WBQualityConstraintsCacheCheckConstraintsResults is
enabled, which currently defaults to false, since it’s not very useful
to store check results in the cache when there’s no code to read them
from there yet.

Bug: T182105
Change-Id: I2fb05b91170eeb08571086fbb3bdc92f88a35015