Read constraint check results from cache
CachingResultsBuilder is updated to read constraint check results from
the object cache if possible (assuming the request does not specify any
statement IDs or constraint IDs). For entities which are not found in
the cache, or where the cached results are outdated, it still does the
regular checking (which includes storing the results in the cache).

Bug: T182106
Change-Id: I9f114fd79cf1201d45f5af47d731084b7e326ce0