Remove unused StatementChangeOpFactory in CheckConstraints
It was apparently unused even back in commit 517b7f720b (change
Icf4ec91bc0) where it was introduced (perhaps a leftover from an earlier
patch set of that change), and then commit 272d7adc74 (change
If5f93c6bf6) removed the declaration of the field, but not its
initialization in the constructor. This commit cleans that up.

Change-Id: I24fb8079a5651bbe9c19a78784b50960ce70863a