Move source code into includes/
The Api and Specials namespaces are moved from their own source
directories into the general includes/ directory, and the api/ and
specicals/ directories are removed.  WikibaseQualityConstraintsHooks.php
is also moved into includes/ according to its namespace.

The Maintenance namespace stays in maintenance/, but is also removed
from the AutoloadNamespaces, because it never needs to be autoloaded
anyways.

Change-Id: I85f4f333a96051d50acc53bbe2a0b545bb1f9e01