Make sure Metadata objects work with assertEquals
Some tests added in later changes use assertEquals() on Metadata
objects, which could previously failed because, for instance,

    Metadata::blank()
    Metadata::ofDependencyMetadata( DependencyMetadata::blank() )

are equivalent but not equal according to assertEquals(). To fix this,
remove the optimization of storing null instead of blank/fresh metadata
objects (which also simplifies some code), so that there is only one
representation of any metadata.

Change-Id: I60a921ba6527336e745644fddc787875fb7faf1b