Simplify the microtime() mocking in CachingResultsBuilderTest
This is a direct follow up for all comments I wrote in I9f114fd, and
should be considered part of T182106.

The most relevant change is the still fragile microtime mocking (hence
the title of this commit). I'm also reintroducing the "makeKey" name,
and use trivial test implementations instead of mocks when possible.

Lucas, feel free to pick this patch up and continue working on it as you
like. If you disagree with certain changes I propose here, feel free to
revert them.

Bug: T182106
Change-Id: I456e6552ccbd1398f5e8fa978a20b0bb2be1109e