Skip constraint checks for qualifiers of certain properties
Properties like Wikidata’s “property example” change the roles of
statement values and qualifiers, which results in lots of pointless
constraint violations that can’t (and shouldn’t) be fixed. Ignore all
qualifiers of such properties.

We don’t add the Wikidata properties “Wikidata property example” and
“Wikidata property example for properties” as the default value, since
there is no way to specify an array config variable that gets
*overwritten* by LocalSettings.php instead of merged with the default
value. Instead, those Wikidata properties are ignored in the
Wikidata-specific configuration in Ica42f7d125. (This also means that
the import snippet in README.md will not import these properties into a
local installation, which is what we want – they’re not required for
WikibaseQualityConstraints to work.)

Bug: T183267
Change-Id: If068c786779122d4f5ff158c9ac8a9a6e6610535