Remove stray @package tag
This file was added in commit a4c1832358 (change I60ad838dda), which was
authored before commit 0d5c5e7156 (change Ib0a0cb771d) (which removed
redundant @package tags) but merged after it, so the @package tag was
accidentally left in the repository. Let’s get rid of it.

Change-Id: I1a711148adf887ca370185d886a31b097bb54e0d