Track cache hit/miss counts per-entity
Increment counters for how many per-entity results we got from the
cache, and how many we didn’t get.

Per-statement results are ignored, since we don’t cache results per
statement. This means that we might increment the counter by zero if
constraint checks for only statement IDs are requested – I verified
manually that statsd is okay with such a non-increment.

Unfortunately, this means that some tests need to switch from skipping
the constructor to providing proper constructor arguments so that we can
inject a NullStatsdDataFactory (instead of the default null value),
which bloats these tests to some degree.

Bug: T184062
Change-Id: I6375603af2f865cc2e881fec91df0488173f5528