Improve ConstraintParameterParserTest function names
The pattern of all the test functions here is “test”, then the name of
the function under test (“parse*Parameter”), and then optionally some
kind of qualifying detail. To improve readability of these names, add
underscores between the name of the tested function and the details, and
within some of the details as well.

Change-Id: I2a74c79265fae1286a5e7f2eee2108e6d59e4cf2