Make constraint report border flush with help link
Several refactorings in ConstraintReportPanel. The main visible effect
is that the border between constraint report panels (added by the
gadget’s style sheet) is now flush on the right side with the “help”
text, since the margin is on the whole panel (including the border)
instead of only on the help text.

In preparation for I5814ef1866, the help link is also wrapped inside a
container, and it’s this whole container which floats to the right (and
has a margin, etc.), not just the individual help link.

And finally, the heading gains a class, instead of being identified in
the style sheet as “the first heading inside the panel”.

Bug: T164351
Change-Id: I548b6072191fc5a06bff06858026062047d7a004