Add $statuses parameter to ResultsBuilder::getResults
This parameter filters the check results represented in the final result
to only those listed in the $statuses. CheckingResultsBuilder implements
this filtering, and CachingResultsBuilder caches only those results
where the $statuses are exactly those three statuses which are also
displayed in the gadget.

For this commit, CheckConstraints passes the list of all statuses to
ResultsBuilder::getResults(). The next commit will add a dedicated API
parameter instead which CheckConstraints just passes through.

Bug: T183927
Change-Id: I3ff5f8d6b0a16a4aa2417b0f1e03b8634376a6dd