Fix status parameter documentation
Originally, I had planned to support caching for requests with e. g.
only the violation|warning statuses as well. However, while implementing
this, I realized that removing results from the response array structure
is not trivial, and so I changed my mind and we now only cache results
for requests which specify exactly the right three statuses.
Unfortunately, I failed to update the commit message to account for
this change.

Bug: T183927
Change-Id: Id81382c35aa5cac5a995432673727bc53bff9eec