Add support for parsing “instance or subclass of” relation
“instance or subclass of” is a new value for the “relation” parameter,
in addition to the existing “instance of” and “subclass of” values.  A
new config variable is added for the item ID of the “instance or
subclass of” relation item, ConstraintParameterParser learns to parse it
(returning the string “relationOrSubclass” – without spaces, since it
will be used in an i18n message key), and the ConstraintParameters test
trait learns to produce the parameter from that string.

Bug: T169858
Change-Id: I0d2a663e5f2200e71640942dab1d21d92eb9d5f7