Add Context::getSnakGroup method
The “snak group” concept could be used by several checkers, but for now,
we will start with the “single value” and “multiple values” checkers.
This is also why we copy the ValueCountCheckerHelper semantics of
ignoring deprecated statements in MainSnakContext (though we might make
this configurable with a method parameter later).

Bug: T185209
Change-Id: I6b926bfe4791b2351c16143b478e06980f03b8f7