Rewrite ValueCountChecker tests
The tests for SingleValueChecker and MultiValueChecker used to load
entities from JSON files and then additionally construct the Statement
for the context independently of them, which was risky because the
constructed Statement was not entirely equivalent to the deserialized
one (its GUID was never set). Since this pattern breaks with the
upcoming changes to ValueCountCheckerHelper (change I301400edd2, for
T175566), rewrite the tests to use NewItem and NewStatement instead.

And on that opportunity, also do some other cleanups: most importantly,
create the mock Constraint object in the constructor, since we never
passed anything other than [] into the constraint parameters anyways
(and there was also a typo on the mocked method name – the real method
name is getConstraintParameters, plural), and also skip the mocking of
methods (never called anyways).

Note: Q1.json and Q6.json are used by UniqueValueCheckerTest and
therefore can’t be deleted yet.

Bug: T168240
Change-Id: I0fced666d5e49d7c7cada49eab19bff972214c9b