Rewrite ValueCountCheckerHelper
ValueCountCheckerHelper now receives a Snak[] instead of a
StatementList, and also receives the property ID to look for instead of
returning an array of the counts of all properties. SingleValueChecker
and MultiValueChecker are adjusted accordingly and pass the snak group
of the context into ValueCountCheckerHelper.

This does not yet provide proper support for checking those constraints
on qualifiers and references, since the checkers declare to not support
those types yet. getSupportedContextTypes will be adjusted in a separate
commit which will also add tests for checking these constraints on
qualifiers and references.

Bug: T175566
Change-Id: I301400edd29b0a10ffc131fd5a20bf6e3b193852