Support qualifier and reference in value count checkers
On the part of the checkers, this just means declaring support for the
'qualifier' and 'reference' context types – all the actual work is done
by the getSnakGroup() implementations and the ValueCountCheckerHelper.
For the tests, I added just one test for qualifiers and one for
references (instead of having tests with one and two snaks for each
context), since I think it’s unlikely that extra tests would uncover any
more faults (both the getSnakGroup() implementations and
ValueCountCheckerHelper already have more tests elsewhere).

Bug: T175566
Change-Id: Id1ef2d371a479f9520c0b135abc0abc01169de09