Add tests for ConstraintParameterRenderer
This adds tests for all the methods of ConstraintParameterRenderer which
are called by constraint checkers or ConstraintParameterParser. On the
other hand, the formatValue and formatParameters methods are only used
for the special page (and by the API if the “detail” and “detailHTML”
fields are included), and might be removed completely depending on the
outcome of T185091.

Bug: T169121
Change-Id: Ifb48530ac0fe26ec1866a8c766c7198a5853d9bc