Rename @dataProvider methods to provide…()
According to MediaWiki guidelines [1], the name of a data provider
should begin with “provide”. Previously, we used several styles
inconsistently in this extension (mostly “get…Provider”). This commit
brings them all in line with the MediaWiki guidelines. In my opinion,
this sometimes makes the tests more readable as well, since the method
name now documents what it actually provides, not just what the test
method to which it belongs is called.

[1]: https://www.mediawiki.org/wiki/Special:PermanentLink/2678449#Data_providers

Bug: T168240
Change-Id: I51f476ff74cf2abd4934f2568268b429f0c39651