Update php-coveralls
There are two things going on here. Firstly, the binary was renamed from
coveralls to php-coveralls, which we need to account for in our Travis
script. However, this binary cannot actually be found in the
satooshi/php-coveralls package – composer prints a warning about that.
I’m not sure what the php-coveralls folks have done here, but in any
event, using the new package name php-coveralls/php-coveralls seems to
fix the problem (it was apparently renamed at some point as well).

Bug: T165617
Change-Id: I7ee5100d447176acbeeae3c50af5c304ef688e55