Make maintenance script testable
Use Maintenance::getConfig(), which has a corresponding setter which
tests can use, instead of getting the config from the services directly.
Move code to get the PropertyInfoLookup, and to instantiate the
UpdateConstraintsTableJob, into the constructor, where it is stored in
private member variables which tests can then later override (using
TestingAccessWrapper to get around the access restrictions). And to make
sure we can actually find the class in the tests, register its
containing directory in the autoload namespaces.

Also removes the use of Maintenance::error()’s deprecated $die
parameter. There is a fatalError() function we could use instead, but in
that case we could completely terminate the tests, which isn’t what we
want. Instead, simply return from execute().

Change-Id: I57f5a102f542db0b59347ec82fe85277ffb95560