Add ViolationMessageRenderer
This will take over the role of ConstraintParameterRenderer in the
future, collecting all the services that are needed to format constraint
violation messages (EntityIdFormatter, ValueFormatter, …). For now, we
start with a simple class with no members that can only format messages
without parameters (since we don’t yet have any parameters in
ViolationMessage either).

Bug: T185710
Change-Id: I55870c43525a88e87c0bcd0d993df4b0ba09f9b0