Accept ViolationMessage in CheckResult
Swapping out all uses of pre-rendered violation messages for
ViolationMessage objects is too much work for one commit, so for a short
time we’ll support both in CheckResult. To simplify use of this hybrid
message member, ViolationMessageRenderer temporarily supports strings as
well, so that you can just take what CheckResult::getMessage() returns
and (if it’s not null) stick it into ViolationMessageRenderer without
having to care whether it’s a string or a ViolationMessage object.

This also documents that a ViolationMessage may be null, which will be
used instead of the empty string if there is no message. However, the
default message (and the one which many checkers explicitly supply in
that case) is still '' – this will be updated in a separate commit.

Bug: T185710
Change-Id: I9bc321afdfb6f1c667febb215066fefc4f86e589