Add support for entity ID arguments to ViolationMessage
A ViolationMessage can now hold a list of arguments, and the first
argument type supported is a single entity ID.
ViolationMessageRenderer gains an EntityIdFormatter and learns to render
a entity ID parameter into the message.

This is intended to replace ConstraintParameterRenderer’s
formatEntityId, formatItemId and formatPropertyId. Note that the latter
two functions accept not only an EntityId object, but also the entity ID
serialization as a string, and will even insert the plain string
(HTML-escaped) if it can’t be parsed as an entity ID. I don’t plan to
replicate this functionality here – it feels unnecessarily risky – but
we need to watch out for this difference when migrating usages.

Bug: T185710
Change-Id: I1e1c8f1703644adb96dfc8a1a937c0943e3128b6