Use ViolationMessage for messages with entity IDs
For messages with no arguments other than entity IDs, we can now use
ViolationMessage with ItemId or PropertyId arguments.

In a few cases, we need to explicitly turn the serialization string into
an ItemId, which the ConstraintParameterRenderer previously did for us.
However, in all of these cases, the serialization should be valid. (Some
tests need to be adjusted to return a valid item ID serialization from
getConstraintTypeItemId().)

Bug: T185710
Change-Id: I212a43ba110a35b987bcf31a73cb99596baea572