Add support for entity ID lists to ViolationMessage
In addition to individual entity IDs, a ViolationMessage can now also
hold a list of entity IDs as a single argument. ViolationMessageRenderer
turns such a list into several Message params: the number of entity IDs,
an HTML list of all formatted entity IDs, and the individual formatted
entity IDs. This follows the behavior of
ConstraintParameterRenderer::formatItemIdList (and the other format…List
methods), though ConstraintParameterRenderer does not include the count
(this is left to the checkers instead).

There is one small difference to ConstraintParameterRenderer in the
truncation behavior: if there are too many entries to list, then a ‘...’
entry is appended to the HTML list, but it is not made a regular message
argument, and the count passed to the message is the truncated count,
not the untruncated one. I think this actually makes more sense. (We
also now use the ellipsis message instead of hard-coding the ‘...’.)

Bug: T185710
Change-Id: I4ce199416261e5ae313470d9d3aa5e1742f7a1a2