Use ViolationMessage for messages with entity ID lists
Now we can also use ViolationMessage for messages with individual entity
ID or entity ID list arguments.

As before, this requires a bit of work on the lists to make them comply
to the stricter requirements. For TypeCheckerHelper, we know it’s safe
to feed the classes into the ItemId constructor because
ConstraintParameterParser::parseClassParameter() already parsed them
once and then extracted only the serialization. For UniqueValueChecker,
we need to watch out because SparqlHelper may also have returned null
elements, which we strip away with an array_filter().

Bug: T185710
Change-Id: Ib6d7a16d93ccd5495c7b4ccffd1fc6843fef46fb