Refactor ViolationMessageRenderer
Most of the code in renderEntityIdList() is extracted into a new
function renderList(), with the function used to render individual
elements (here: renderEntityId()) as a parameter. This will be useful
for rendering ItemIdSnakValue lists.

Most of the tests for renderEntityIdList() are also changed to be tests
for the generic renderList(), with only two general tests remaining to
test renderEntityIdList() specifically.

Bug: T185710
Change-Id: I04aced42ae7b3f41f4f2f965561e8cef6c7aa34e