Use ViolationMessage for even more checkers
Plenty of checkers were using wfMessage() in an exceptional case where a
data value has the wrong type – with data value type argument support in
ViolationMessage, they can now use ViolationMessage for that as well.

(This also resolves the TODOs about lacking a message for the
wikibase-entity data value type, which is now handled in
ViolationMessageRenderer – see previous commit.)

Some tests need to be updated to return a valid item ID serialization in
a constraint mock’s getConstraintTypeItemId() method.

Bug: T185710
Change-Id: If25205bb180512280182bc603dd89c19cb2ca2c8