Inject ValueFormatter into ViolationMessageRenderer
ViolationMessageRenderer will soon need a ValueFormatter to render
DataValue arguments, and the injection touches enough places that I
think it makes sense to do it in a separate commit, even if it briefly
looks like an unused field.

In the tests, UnDeserializableValueFormatter is used as the best analog
of PlainEntityIdFormatter that I could find.
ViolationMessageRendererTest, which creates ViolationMessageRenderers in
a lot of different places, gets a new helper method to create a
ViolationMessageRenderer with default arguments so that the two long
class names of the default implementations don’t have to be repeated
everywhere.

Change-Id: I0da2dcbdab5e52e526ac68df75d4664d011a3108