Only compute difference in years on time values
DiffWithinRangeChecker special-cases the unit “year” in order to support
specifying that, for instance, the difference between a person’s date of
birth and date of death should be within 0 to 150 years. However, we
should only apply that special case if the difference is actually
between two time values – if the two values are actually quantities,
attempting to compute the difference in years between them will result
in a TypeError.

Bug: T186226
Change-Id: I127937bbb798d87049e8ae8370c1de9c94ad9e99