Add support for multilingual text to ViolationMessage
Multilingual text arguments are used in the violation message of the
“format” constraint (for the syntax clarification), but this message
behaves in a very special way – if the multilingual text is not
available in the user’s language, a different message key is used,
referring to a version of the message which does not include the syntax
clarification parameter.

This becomes tricky to support in ViolationMessageRenderer, so the logic
to handle this is kept in a separate class instead, which knows which
messages take such an argument (currently, only one) and delegates
rendering of all other messages back to the regular
ViolationMessageRenderer.

The tests are mostly based on the ConstraintParameterParser tests, since
ConstraintParameterParser also implements such a fallback logic (it will
be removed in the next change).

Bug: T185710
Change-Id: Idf2b7de58b3979a1ed3f7c5a0368041f4176f4ac