Use ViolationMessage for FormatChecker
ConstraintParameterParser::parseSyntaxClarificationParameter() is
changed to return a MultilingualTextValue instead of picking the best
text for the current language already, and FormatChecker passes that
into the ViolationMessage. This means that we can remove a lot of
fallback-related tests from ConstraintParameterParserTest (they are now
in ViolationMessageRendererTest instead, added there in the previous
commit).

As a side effect, FormatChecker tests now need to expect the message key
version with clarification, since that is considered the “main” message,
and the version without clarification is just a fallback if the
clarification is not available in the user’s language.

Bug: T185710
Change-Id: I862509dc4ff3d80f96a5d5c08f7caa9461d8fa7a