Inject ViolationMessageSerializer and -Deserializer
SparqlHelper serializes and deserializes ConstraintParameterExceptions
in the regex cache, so once ConstraintParameterExceptions contain
ViolationMessages instead of plain strings, it will need a
ViolationMessageSerializer and a ViolationMessageDeserializer. This
commit adds them to SparqlHelper and also ConstraintReportFactory.

In SparqlHelperTest, a helper method to create SparqlHelpers is
extracted, since many tests used the same copied instantiation.

Bug: T187186
Change-Id: I8c6e4cd29755496e87766ef51faab1d6d62021cd